weatherstation/requirements.txt
Django==1.7.6
djangorestframework==3.1.0
wsgiref==0.1.2
